Restructured parser
parser was restructured to fix parsing on windows. Using pyparsing,
windows could not identify #EQUATIONS section if #EQUATIONS was not
followed by another section. This could be solved by changing the
multi-regular expression style.

However, this led to a redesign that is simpler and more robust. I am
using regular expressions to find sections and pyparsing just for
reactions.